+++ package/Runtime/Matrix2D.cs
@@ -284,7 +284,24 @@ namespace Unity.VectorGraphics
         /// <summary>Creates a rotation matrix. Note: Assumes unit quaternion.</summary>
         /// <param name="angleRadians">The rotation angle, in radians</param>
         /// <returns>The rotation matrix</returns>
+        [Obsolete("Use RotateRH or RotateLH instead")]
         public static Matrix2D Rotate(float angleRadians)
+        {
+            return RotateLH(angleRadians);
+        }
+
+        /// <summary>Creates a right-hand side rotation matrix.</summary>
+        /// <param name="angleRadians">The rotation angle, in radians</param>
+        /// <returns>The rotation matrix</returns>
+        public static Matrix2D RotateRH(float angleRadians)
+        {
+            return RotateLH(-angleRadians);
+        }
+
+        /// <summary>Creates a left-hand side rotation matrix.</summary>
+        /// <param name="angleRadians">The rotation angle, in radians</param>
+        /// <returns>The rotation matrix</returns>
+        public static Matrix2D RotateLH(float angleRadians)
         {
             // No SinCos? I hope the compiler optimizes this
             float s = Mathf.Sin(angleRadians);